Refactoring: Split angular cli flags and args into separate files Also split command definition and handlers appart
