+++ src/angular/angular-cli-arguments.ts
@@ -1,0 +1,19 @@
+export interface IAngularCliArguments {
+    aot?: boolean;
+    app?: boolean;
+    "deploy-url"?: string;
+    environment?: string;
+    "extract-css"?: boolean;
+    force?: boolean;
+    "i18n-file"?: string;
+    "i18n-format"?: string;
+    locale?: string;
+    "missing-translation"?: string;
+    "output-hashing"?: string;
+    progress?: boolean;
+    sourcemap?: boolean;
+    target?: string;
+    "vendor-chunk"?: boolean;
+    "common-chunk"?: boolean;
+    verbose?: boolean;
+}
