+++ src/angular/angular-cli-flags.ts
@@ -1,0 +1,19 @@
+export const ANGULAR_CLI_FLAGS: string[] = [
+    "aot",
+    "app",
+    "common-chunk",
+    "deploy-url",
+    "environment",
+    "extract-css",
+    "force",
+    "i18n-file",
+    "i18n-format",
+    "locale",
+    "missing-translation",
+    "output-hashing",
+    "progress",
+    "sourcemap",
+    "target",
+    "vendor-chunk",
+    "verbose",
+];
