+++ src/commands/serve.ts
@@ -1,0 +1,9 @@
+import { ServeTask } from "../tasks/serve";
+
+export const serveCommand = {
+    command: "serve",
+    describe: "Serve the app in Electron window",
+    handler: (argv) => {
+        ServeTask.serve();
+    },
+};
